collector/requirements.txt
requests==2.32.4
psutil==7.0.0
python-dotenv==1.1.1